+++ src/main/java/org/mariadb/jdbc/Configuration.java
@@ -63,6 +63,7 @@ public class Configuration {
 
   // various
   private String timezone = null;
+  private String connectionCollation = null;
   private String connectionTimeZone = null;
   private Boolean forceConnectionTimeZoneToSession = null;
   private boolean preserveInstants;
@@ -105,7 +106,9 @@ public class Configuration {
   private SslMode sslMode = SslMode.DISABLE;
   private String serverSslCert = null;
   private String keyStore = null;
+  private String trustStore = null;
   private String keyStorePassword = null;
+  private String trustStorePassword = null;
   private String keyPassword = null;
   private String keyStoreType = null;
   private String trustStoreType = null;
@@ -174,6 +177,7 @@ public class Configuration {
       Properties nonMappedOptions,
       String timezone,
       String connectionTimeZone,
+      String connectionCollation,
       boolean forceConnectionTimeZoneToSession,
       boolean preserveInstants,
       Boolean autocommit,
@@ -210,7 +214,9 @@ public class Configuration {
       SslMode sslMode,
       String serverSslCert,
       String keyStore,
+      String trustStore,
       String keyStorePassword,
+      String trustStorePassword,
       String keyPassword,
       String keyStoreType,
       String trustStoreType,
@@ -261,6 +267,7 @@ public class Configuration {
     this.nonMappedOptions = nonMappedOptions;
     this.timezone = timezone;
     this.connectionTimeZone = connectionTimeZone;
+    this.connectionCollation = connectionCollation;
     this.forceConnectionTimeZoneToSession = forceConnectionTimeZoneToSession;
     this.preserveInstants = preserveInstants;
     this.autocommit = autocommit;
@@ -297,7 +304,9 @@ public class Configuration {
     this.sslMode = sslMode;
     this.serverSslCert = serverSslCert;
     this.keyStore = keyStore;
+    this.trustStore = trustStore;
     this.keyStorePassword = keyStorePassword;
+    this.trustStorePassword = trustStorePassword;
     this.keyPassword = keyPassword;
     this.keyStoreType = keyStoreType;
     this.trustStoreType = trustStoreType;
@@ -378,6 +387,7 @@ public class Configuration {
       Boolean yearIsDateType,
       String timezone,
       String connectionTimeZone,
+      String connectionCollation,
       Boolean forceConnectionTimeZoneToSession,
       Boolean preserveInstants,
       Boolean dumpQueriesOnException,
@@ -419,7 +429,9 @@ public class Configuration {
       Boolean allowPublicKeyRetrieval,
       String serverSslCert,
       String keyStore,
+      String trustStore,
       String keyStorePassword,
+      String trustStorePassword,
       String keyPassword,
       String keyStoreType,
       String trustStoreType,
@@ -475,6 +487,7 @@ public class Configuration {
     if (yearIsDateType != null) this.yearIsDateType = yearIsDateType;
     this.timezone = timezone;
     if (connectionTimeZone != null) this.connectionTimeZone = connectionTimeZone;
+    if (connectionCollation != null) this.connectionCollation = connectionCollation;
     if (forceConnectionTimeZoneToSession != null)
       this.forceConnectionTimeZoneToSession = forceConnectionTimeZoneToSession;
     if (preserveInstants != null) this.preserveInstants = preserveInstants;
@@ -555,7 +568,9 @@ public class Configuration {
     if (initSql != null) this.initSql = initSql;
     if (serverSslCert != null) this.serverSslCert = serverSslCert;
     if (keyStore != null) this.keyStore = keyStore;
+    if (trustStore != null) this.trustStore = trustStore;
     if (keyStorePassword != null) this.keyStorePassword = keyStorePassword;
+    if (trustStorePassword != null) this.trustStorePassword = trustStorePassword;
     if (keyPassword != null) this.keyPassword = keyPassword;
     if (keyStoreType != null) this.keyStoreType = keyStoreType;
     if (trustStoreType != null) this.trustStoreType = trustStoreType;
@@ -618,6 +633,24 @@ public class Configuration {
     // option value verification
     // *************************************************************
 
+    // ensure connection collation format
+    if (connectionCollation != null) {
+      if ("".equals(connectionCollation.trim())) {
+        this.connectionCollation = null;
+      } else {
+        // ensure this is an utf8mb4 collation
+        if (!connectionCollation.toLowerCase(Locale.ROOT).startsWith("utf8mb4_")) {
+          throw new SQLException(
+              String.format(
+                  "wrong connection collation '%s' only utf8mb4 collation are accepted",
+                  connectionCollation));
+        } else if (!connectionCollation.matches("^[a-zA-Z0-9_]+$")) {
+          throw new SQLException(
+              String.format("wrong connection collation '%s' name", connectionCollation));
+        }
+      }
+    }
+
     // int fields must all be positive
     Field[] fields = Configuration.class.getDeclaredFields();
     try {
@@ -651,6 +684,7 @@ public class Configuration {
             .haMode(this.haMode)
             .timezone(this.timezone)
             .connectionTimeZone(this.connectionTimeZone)
+            .connectionCollation(this.connectionCollation)
             .forceConnectionTimeZoneToSession(this.forceConnectionTimeZoneToSession)
             .preserveInstants(this.preserveInstants)
             .autocommit(this.autocommit)
@@ -689,8 +723,10 @@ public class Configuration {
             .sslMode(this.sslMode.name())
             .serverSslCert(this.serverSslCert)
             .keyStore(this.keyStore)
+            .trustStore(this.trustStore)
             .keyStoreType(this.keyStoreType)
             .keyStorePassword(this.keyStorePassword)
+            .trustStorePassword(this.trustStorePassword)
             .keyPassword(this.keyPassword)
             .trustStoreType(this.trustStoreType)
             .enabledSslCipherSuites(this.enabledSslCipherSuites)
@@ -799,7 +835,12 @@ public class Configuration {
       int skipPos;
       int posToSkip = 0;
       while ((skipPos = urlSecondPart.indexOf("address=(", posToSkip)) > -1) {
-        posToSkip = urlSecondPart.indexOf(")", skipPos);
+        posToSkip = urlSecondPart.indexOf(")", skipPos) + 1;
+        while (urlSecondPart.startsWith("(", posToSkip)) {
+          int endingBraceIndex = urlSecondPart.indexOf(")", posToSkip);
+          if (endingBraceIndex == -1) break;
+          posToSkip = endingBraceIndex + 1;
+        }
       }
       int dbIndex = urlSecondPart.indexOf("/", posToSkip);
       int paramIndex = urlSecondPart.indexOf("?");
@@ -1043,13 +1084,18 @@ public class Configuration {
               case "Integer":
               case "SslMode":
               case "CatalogTerm":
-                (Objects.equals(fieldValue, field.get(defaultConf))
+                StringBuilder sbb =
+                    (Objects.equals(fieldValue, field.get(defaultConf))
                         ? sbDefaultOpts
-                        : sbDifferentOpts)
-                    .append("\n * ")
-                    .append(field.getName())
-                    .append(" : ")
-                    .append(fieldValue);
+                        : sbDifferentOpts);
+
+                sbb.append("\n * ").append(field.getName()).append(" : ");
+                if ("password".equals(field.getName())
+                    || "keyStorePassword".equals(field.getName())
+                    || "trustStorePassword".equals(field.getName())) {
+                  sbb.append("***");
+                } else sbb.append(fieldValue);
+
                 break;
               case "ArrayList":
                 (Objects.equals(fieldValue.toString(), field.get(defaultConf).toString())
@@ -1144,7 +1190,9 @@ public class Configuration {
 
         if (obj != null && (!(obj instanceof Properties) || ((Properties) obj).size() > 0)) {
 
-          if ("password".equals(field.getName())) {
+          if ("password".equals(field.getName())
+              || "keyStorePassword".equals(field.getName())
+              || "trustStorePassword".equals(field.getName())) {
             sb.append(first ? '?' : '&');
             first = false;
             sb.append(field.getName()).append('=');
@@ -1336,12 +1384,30 @@ public class Configuration {
   }
 
   /**
+   * trust store
+   *
+   * @return trust store
+   */
+  public String trustStore() {
+    return trustStore;
+  }
+
+  /**
    * key store password
    *
    * @return key store password
    */
   public String keyStorePassword() {
     return keyStorePassword;
+  }
+
+  /**
+   * trust store password
+   *
+   * @return trust store password
+   */
+  public String trustStorePassword() {
+    return trustStorePassword;
   }
 
   /**
@@ -1633,6 +1699,15 @@ public class Configuration {
    */
   public String connectionTimeZone() {
     return connectionTimeZone;
+  }
+
+  /**
+   * get connectionCollation
+   *
+   * @return connectionCollation
+   */
+  public String connectionCollation() {
+    return connectionCollation;
   }
 
   /**
@@ -2124,6 +2199,7 @@ public class Configuration {
     // various
     private String timezone;
     private String connectionTimeZone;
+    private String connectionCollation;
     private Boolean forceConnectionTimeZoneToSession;
     private Boolean preserveInstants;
     private Boolean autocommit;
@@ -2164,7 +2240,9 @@ public class Configuration {
     private String sslMode;
     private String serverSslCert;
     private String keyStore;
+    private String trustStore;
     private String keyStorePassword;
+    private String trustStorePassword;
     private String keyPassword;
     private String keyStoreType;
     private String trustStoreType;
@@ -2257,6 +2335,18 @@ public class Configuration {
     }
 
     /**
+     * File path of the trustStore file that contain trusted certificates (similar to java System
+     * property \"javax.net.ssl.trustStore\")
+     *
+     * @param trustStore client trust store certificates
+     * @return this {@link Builder}
+     */
+    public Builder trustStore(String trustStore) {
+      this.trustStore = nullOrEmpty(trustStore);
+      return this;
+    }
+
+    /**
      * Client keystore password
      *
      * @param keyStorePassword client store password
@@ -2264,6 +2354,17 @@ public class Configuration {
      */
     public Builder keyStorePassword(String keyStorePassword) {
       this.keyStorePassword = nullOrEmpty(keyStorePassword);
+      return this;
+    }
+
+    /**
+     * Client truststore password
+     *
+     * @param trustStorePassword client truststore password
+     * @return this {@link Builder}
+     */
+    public Builder trustStorePassword(String trustStorePassword) {
+      this.trustStorePassword = nullOrEmpty(trustStorePassword);
       return this;
     }
 
@@ -2800,6 +2901,18 @@ public class Configuration {
      */
     public Builder connectionTimeZone(String connectionTimeZone) {
       this.connectionTimeZone = nullOrEmpty(connectionTimeZone);
+      return this;
+    }
+
+    /**
+     * indicate what utf8mb4 collation to use. if not set, server default collation for utf8mb4 will
+     * be used
+     *
+     * @param connectionCollation utf8mb4 collation to use
+     * @return this {@link Builder}
+     */
+    public Builder connectionCollation(String connectionCollation) {
+      this.connectionCollation = nullOrEmpty(connectionCollation);
       return this;
     }
 
@@ -3345,6 +3458,7 @@ public class Configuration {
               this.yearIsDateType,
               this.timezone,
               this.connectionTimeZone,
+              this.connectionCollation,
               this.forceConnectionTimeZoneToSession,
               this.preserveInstants,
               this.dumpQueriesOnException,
@@ -3386,7 +3500,9 @@ public class Configuration {
               this.allowPublicKeyRetrieval,
               this.serverSslCert,
               this.keyStore,
+              this.trustStore,
               this.keyStorePassword,
+              this.trustStorePassword,
               this.keyPassword,
               this.keyStoreType,
               this.trustStoreType,